requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

langchain-core
langchain-google-genai
langgraph

playwright
requests
pymupdf
pytesseract
Pillow

vnstock
pandas
pandas-ta
openpyxl
dateparser

psycopg2-binary
transformers
torch